+++ lib/board_datetime_picker.dart
@@ -23,7 +23,8 @@ export 'src/board_datetime_widget.dart'
         showBoardDateTimePicker,
         showBoardDateTimePickerForDateTime,
         showBoardDateTimePickerForDate,
-        showBoardDateTimePickerForTime;
+        showBoardDateTimePickerForTime,
+        showBoardDateTimeMultiPicker;
 export 'src/ui/parts/focus_node.dart' show BoardDateTimeInputFocusNode;
 export 'src/utils/board_datetime_result.dart';
 export 'src/utils/board_enum.dart' show DateTimePickerType;